Refactor logging and authorization
